refactor: Add #[must_use] annotations to various functions for better usage clarity

+++ src/text/text_typewriter_ops.rs
@@ -107,7 +107,7 @@ impl TextSystem {
     }
 
     #[must_use]
-    pub fn get_progress(&self, id: usize) -> f32 {
+    pub fn get_progress(&self, id: usize) -> f64 {
         self.typewriter_instance
             .get_effect(id)
             .map_or(0.0, TypewriterEffect::progress)
